requirements.txt
agno
python-dotenv
duckduckgo-search
pypdf
lancedb
streamlit
pandas
tantivy
yfinance
google-generativeai
groq
google-genai
flask